Addresses #1308 Fix suspiscious cast in range tree hash table
git-svn-id: file:///svn/toku/tokudb.1032b@8294 c7de825b-a66e-492c-adef-691d508d4ae1

+++ src/lock_tree/rth.c
@@ -19,7 +19,7 @@
 const u_int32_t __toku_rth_init_size = 521;
 
 static inline u_int32_t toku__rth_hash(toku_rth* rth, TXNID key) {
-    size_t tmp = (size_t)key;
+    u_int64_t tmp = (u_int64_t)key;
     return tmp % rth->num_buckets;
 }
